// src/app/core/models/dashboardWidgetDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */

export interface DashboardWidgetDto {
  id: string;
  name: string;
  type: string;
  configuration: string;
  isEnabled: boolean;
  order: number;
}