// src/app/core/models/roleDto.ts
/**
 * /**
 *  * Generated by orval v1.0.0 🍺
 *  * Do not edit manually.
 *  * Funeral API
 *  * API for managing funeral services, members, policies, claims, and more.
 *  *\/
 */
import type { RoleDtoPermissionsItem } from './roleDtoPermissionsItem';

export interface RoleDto {
  id: string;
  name: string;
  permissions: RoleDtoPermissionsItem[];
}